+++ package.json
@@ -2,19 +2,23 @@
   "name": "baseball",
   "version": "1.0.0",
   "description": "",
-  "main": "./src/index.js",
+  "main": "./src/index.mjs",
   "scripts": {
-    "start": "node ./src/index.js",
+    "start": "node ./src/index.mjs",
     "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon ./src/index.js"
+    "dev": "nodemon ./src/index.mjs"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "apollo-server": "^2.25.3",
+    "@apollo/server": "^4.10.0",
+    "apollo-server": "^3.13.0",
     "dotenv": "^8.6.0",
-    "graphql": "^15.8.0",
+    "graphql": "^16.8.1",
     "mongoose": "^5.13.14"
+  },
+  "devDependencies": {
+    "nodemon": "^3.0.2"
   }
 }
