--- src/Parks/index.js
@@ -1,10 +1,0 @@
-// Park domain
-const { Park } = require("./Park");
-const { resolvers } = require("./resolvers");
-const { typeDef } = require("./typeDef");
-
-module.exports = {
-  Park,
-  resolvers,
-  typeDef,
-};
