+++ src/Parks/index.mjs
@@ -1,0 +1,9 @@
+// Park domain
+// const { Park } = require("./Park.mjs");
+// const { resolvers } = require("./resolvers.mjs");
+// const { typeDef } = require("./typeDef.mjs");
+import Park from "./Park.mjs";
+import resolvers from "./resolvers.mjs";
+import typeDef from "./typeDef.mjs";
+
+export default { Park, resolvers, typeDef };
