--- src/index.js
@@ -1,54 +1,0 @@
-require("dotenv").config();
-const { ApolloServer, gql } = require("apollo-server");
-const connectDB = require("./config/db");
-const appearances = require("./Appearances");
-const batting = require("./Batting");
-const franchises = require("./Franchises");
-const league = require("./Leagues");
-const lookups = require("./Lookups");
-const parks = require("./Parks");
-const picks = require("./Picks");
-const pitching = require("./Pitching");
-const players = require("./Players");
-
-const typeDef = gql`
-  type Query
-  type Mutation
-`;
-
-connectDB();
-
-const server = new ApolloServer({
-  typeDefs: [
-    typeDef,
-    appearances.typeDef,
-    batting.typeDef,
-    franchises.typeDef,
-    league.typeDef,
-    lookups.typeDef,
-    parks.typeDef,
-    picks.typeDef,
-    pitching.typeDef,
-    players.typeDef,
-  ],
-  resolvers: [
-    appearances.resolvers,
-    batting.resolvers,
-    franchises.resolvers,
-    league.resolvers,
-    lookups.resolvers,
-    parks.resolvers,
-    picks.resolvers,
-    pitching.resolvers,
-    players.resolvers,
-  ],
-  playground: true,
-});
-
-// db.on("error", console.error.bind(console, "connection error:"));
-// db.once("open", function () {
-//   console.log("✅ Database connection established ✅");
-server.listen({ port: process.env.PORT || 4000 }).then(({ url }) => {
-  console.log(`Server ready at ${url}`);
-});
-// });
